requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Markdown==3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
